Add: GetCircularTexture function
In this code, a circular crosshair texture is generated using a nested loop to iterate over each pixel of the texture. The distance between each pixel and the center of the texture is calculated using the Vector2.Distance function. If this distance is less than or equal to the radius of the circle, the pixel is set to white, otherwise it is set to transparent. The resulting texture is then drawn on the screen using the GUI.DrawTexture function.